fix(laporan): correct migration DataType, null interface type, strengthen test — L2 review fixes

+++ backend/src/migrations/20260517_alter_laporan_realisasi_rp_nullable.ts
@@ -3,14 +3,14 @@ import { QueryInterface, DataTypes } from 'sequelize';
 export default {
   up: async (queryInterface: QueryInterface) => {
     await queryInterface.changeColumn('laporan', 'realisasi_rp', {
-      type: DataTypes.DECIMAL(18, 2),
+      type: DataTypes.BIGINT,
       allowNull: true,
       defaultValue: null,
     });
   },
   down: async (queryInterface: QueryInterface) => {
     await queryInterface.changeColumn('laporan', 'realisasi_rp', {
-      type: DataTypes.DECIMAL(18, 2),
+      type: DataTypes.BIGINT,
       allowNull: false,
       defaultValue: 0,
     });
